woot; much cleaner rotations

+++ starry2/vectorize.h
@@ -37,7 +37,7 @@ namespace vectorize {
         // 1
         T operator()(const T& arg1) {
             T res(arg1.size());
-            for (long i = 0; i < arg1.size(); i++) res(i) = function(arg1(i));
+            for (size_t i = 0; i < arg1.size(); i++) res(i) = function(arg1(i));
             return res;
         }
 
@@ -65,21 +65,21 @@ namespace vectorize {
         // 01
         T operator()(const Scalar<T>& arg1, const T& arg2) {
             T res(arg2.size());
-            for (long i = 0; i < arg2.size(); i++) res(i) = function(arg1, arg2(i));
+            for (size_t i = 0; i < arg2.size(); i++) res(i) = function(arg1, arg2(i));
             return res;
         }
 
         // 10
         T operator()(const T& arg1, const Scalar<T>& arg2) {
             T res(arg1.size());
-            for (long i = 0; i < arg1.size(); i++) res(i) = function(arg1(i), arg2);
+            for (size_t i = 0; i < arg1.size(); i++) res(i) = function(arg1(i), arg2);
             return res;
         }
 
         // 11
         T operator()(const T& arg1, const T& arg2) {
             T res(arg1.size());
-            for (long i = 0; i < arg1.size(); i++) res(i) = function(arg1(i), arg2(i));
+            for (size_t i = 0; i < arg1.size(); i++) res(i) = function(arg1(i), arg2(i));
             return res;
         }
 
@@ -107,49 +107,49 @@ namespace vectorize {
         // 001
         T operator()(const Scalar<T>& arg1, const Scalar<T>& arg2, const T& arg3) {
             T res(arg3.size());
-            for (long i = 0; i < arg3.size(); i++) res(i) = function(arg1, arg2, arg3(i));
+            for (size_t i = 0; i < arg3.size(); i++) res(i) = function(arg1, arg2, arg3(i));
             return res;
         }
 
         // 010
         T operator()(const Scalar<T>& arg1, const T& arg2, const Scalar<T>& arg3) {
             T res(arg2.size());
-            for (long i = 0; i < arg2.size(); i++) res(i) = function(arg1, arg2(i), arg3);
+            for (size_t i = 0; i < arg2.size(); i++) res(i) = function(arg1, arg2(i), arg3);
             return res;
         }
 
         // 011
         T operator()(const Scalar<T>& arg1, const T& arg2, const T& arg3) {
             T res(arg2.size());
-            for (long i = 0; i < arg2.size(); i++) res(i) = function(arg1, arg2(i), arg3(i));
+            for (size_t i = 0; i < arg2.size(); i++) res(i) = function(arg1, arg2(i), arg3(i));
             return res;
         }
 
         // 100
         T operator()(const T& arg1, const Scalar<T>& arg2, const Scalar<T>& arg3) {
             T res(arg1.size());
-            for (long i = 0; i < arg1.size(); i++) res(i) = function(arg1(i), arg2, arg3);
+            for (size_t i = 0; i < arg1.size(); i++) res(i) = function(arg1(i), arg2, arg3);
             return res;
         }
 
         // 101
         T operator()(const T& arg1, const Scalar<T>& arg2, const T& arg3) {
             T res(arg1.size());
-            for (long i = 0; i < arg1.size(); i++) res(i) = function(arg1(i), arg2, arg3(i));
+            for (size_t i = 0; i < arg1.size(); i++) res(i) = function(arg1(i), arg2, arg3(i));
             return res;
         }
 
         // 110
         T operator()(const T& arg1, const T& arg2, const Scalar<T>& arg3) {
             T res(arg1.size());
-            for (long i = 0; i < arg1.size(); i++) res(i) = function(arg1(i), arg2(i), arg3);
+            for (size_t i = 0; i < arg1.size(); i++) res(i) = function(arg1(i), arg2(i), arg3);
             return res;
         }
 
         // 111
         T operator()(const T& arg1, const T& arg2, const T& arg3) {
             T res(arg1.size());
-            for (long i = 0; i < arg1.size(); i++) res(i) = function(arg1(i), arg2(i), arg3(i));
+            for (size_t i = 0; i < arg1.size(); i++) res(i) = function(arg1(i), arg2(i), arg3(i));
             return res;
         }
 
@@ -177,105 +177,105 @@ namespace vectorize {
         // 0001
         T operator()(const Scalar<T>& arg1, const Scalar<T>& arg2, const Scalar<T>& arg3, const T& arg4) {
             T res(arg4.size());
-            for (long i = 0; i < arg4.size(); i++) res(i) = function(arg1, arg2, arg3, arg4(i));
+            for (size_t i = 0; i < arg4.size(); i++) res(i) = function(arg1, arg2, arg3, arg4(i));
             return res;
         }
 
         // 0010
         T operator()(const Scalar<T>& arg1, const Scalar<T>& arg2, const T& arg3, const Scalar<T>& arg4) {
             T res(arg3.size());
-            for (long i = 0; i < arg3.size(); i++) res(i) = function(arg1, arg2, arg3(i), arg4);
+            for (size_t i = 0; i < arg3.size(); i++) res(i) = function(arg1, arg2, arg3(i), arg4);
             return res;
         }
 
         // 0011
         T operator()(const Scalar<T>& arg1, const Scalar<T>& arg2, const T& arg3, const T& arg4) {
             T res(arg3.size());
-            for (long i = 0; i < arg3.size(); i++) res(i) = function(arg1, arg2, arg3(i), arg4(i));
+            for (size_t i = 0; i < arg3.size(); i++) res(i) = function(arg1, arg2, arg3(i), arg4(i));
             return res;
         }
 
         // 0100
         T operator()(const Scalar<T>& arg1, const T& arg2, const Scalar<T>& arg3, const Scalar<T>& arg4) {
             T res(arg2.size());
-            for (long i = 0; i < arg2.size(); i++) res(i) = function(arg1, arg2(i), arg3, arg4);
+            for (size_t i = 0; i < arg2.size(); i++) res(i) = function(arg1, arg2(i), arg3, arg4);
             return res;
         }
 
         // 0101
         T operator()(const Scalar<T>& arg1, const T& arg2, const Scalar<T>& arg3, const T& arg4) {
             T res(arg2.size());
-            for (long i = 0; i < arg2.size(); i++) res(i) = function(arg1, arg2(i), arg3, arg4(i));
+            for (size_t i = 0; i < arg2.size(); i++) res(i) = function(arg1, arg2(i), arg3, arg4(i));
             return res;
         }
 
         // 0110
         T operator()(const Scalar<T>& arg1, const T& arg2, const T& arg3, const Scalar<T>& arg4) {
             T res(arg2.size());
-            for (long i = 0; i < arg2.size(); i++) res(i) = function(arg1, arg2(i), arg3(i), arg4);
+            for (size_t i = 0; i < arg2.size(); i++) res(i) = function(arg1, arg2(i), arg3(i), arg4);
             return res;
         }
 
         // 0111
         T operator()(const Scalar<T>& arg1, const T& arg2, const T& arg3, const T& arg4) {
             T res(arg2.size());
-            for (long i = 0; i < arg2.size(); i++) res(i) = function(arg1, arg2(i), arg3(i), arg4(i));
+            for (size_t i = 0; i < arg2.size(); i++) res(i) = function(arg1, arg2(i), arg3(i), arg4(i));
             return res;
         }
 
         // 1000
         T operator()(const T& arg1, const Scalar<T>& arg2, const Scalar<T>& arg3, const Scalar<T>& arg4) {
             T res(arg1.size());
-            for (long i = 0; i < arg1.size(); i++) res(i) = function(arg1(i), arg2, arg3, arg4);
+            for (size_t i = 0; i < arg1.size(); i++) res(i) = function(arg1(i), arg2, arg3, arg4);
             return res;
         }
 
         // 0001
         T operator()(const T& arg1, const Scalar<T>& arg2, const Scalar<T>& arg3, const T& arg4) {
             T res(arg1.size());
-            for (long i = 0; i < arg1.size(); i++) res(i) = function(arg1(i), arg2, arg3, arg4(i));
+            for (size_t i = 0; i < arg1.size(); i++) res(i) = function(arg1(i), arg2, arg3, arg4(i));
             return res;
         }
 
         // 0010
         T operator()(const T& arg1, const Scalar<T>& arg2, const T& arg3, const Scalar<T>& arg4) {
             T res(arg1.size());
-            for (long i = 0; i < arg1.size(); i++) res(i) = function(arg1(i), arg2, arg3(i), arg4);
+            for (size_t i = 0; i < arg1.size(); i++) res(i) = function(arg1(i), arg2, arg3(i), arg4);
             return res;
         }
 
         // 0011
         T operator()(const T& arg1, const Scalar<T>& arg2, const T& arg3, const T& arg4) {
             T res(arg1.size());
-            for (long i = 0; i < arg1.size(); i++) res(i) = function(arg1(i), arg2, arg3(i), arg4(i));
+            for (size_t i = 0; i < arg1.size(); i++) res(i) = function(arg1(i), arg2, arg3(i), arg4(i));
             return res;
         }
 
         // 0100
         T operator()(const T& arg1, const T& arg2, const Scalar<T>& arg3, const Scalar<T>& arg4) {
             T res(arg1.size());
-            for (long i = 0; i < arg1.size(); i++) res(i) = function(arg1(i), arg2(i), arg3, arg4);
+            for (size_t i = 0; i < arg1.size(); i++) res(i) = function(arg1(i), arg2(i), arg3, arg4);
             return res;
         }
 
         // 0101
         T operator()(const T& arg1, const T& arg2, const Scalar<T>& arg3, const T& arg4) {
             T res(arg1.size());
-            for (long i = 0; i < arg1.size(); i++) res(i) = function(arg1(i), arg2(i), arg3, arg4(i));
+            for (size_t i = 0; i < arg1.size(); i++) res(i) = function(arg1(i), arg2(i), arg3, arg4(i));
             return res;
         }
 
         // 0110
         T operator()(const T& arg1, const T& arg2, const T& arg3, const Scalar<T>& arg4) {
             T res(arg1.size());
-            for (long i = 0; i < arg1.size(); i++) res(i) = function(arg1(i), arg2(i), arg3(i), arg4);
+            for (size_t i = 0; i < arg1.size(); i++) res(i) = function(arg1(i), arg2(i), arg3(i), arg4);
             return res;
         }
 
         // 0111
         T operator()(const T& arg1, const T& arg2, const T& arg3, const T& arg4) {
             T res(arg1.size());
-            for (long i = 0; i < arg1.size(); i++) res(i) = function(arg1(i), arg2(i), arg3(i), arg4(i));
+            for (size_t i = 0; i < arg1.size(); i++) res(i) = function(arg1(i), arg2(i), arg3(i), arg4(i));
             return res;
         }
